Move Cello.h one directory deeper
