--- examples/help.c
@@ -1,9 +1,0 @@
-#include "Cello.h"
-
-int main(int argc, char** argv) {
-  
-  help(Range);
-  
-  return 0;
-}
-
