+++ examples/src/cello_world.c
@@ -1,0 +1,6 @@
+#include <Cello/Cello.h>
+
+int main(int argc, char** argv) {
+  println("Cello World!");
+  return 0;
+}
