+++ examples/src/help.c
@@ -1,6 +1,9 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 int main(int argc, char** argv) {
-  println("Cello World!");
+  
+  help(Range);
+  
   return 0;
 }
+
