+++ examples/src/object.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 /* Define a normal C structure */
 struct Point {
