+++ src/Array.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Array_Name(void) {
   return "Array";
