+++ src/Assign.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Assign_Name(void) {
   return "Assign";
