+++ src/Cmp.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Cmp_Name(void) {
   return "Cmp";
