+++ src/Concat.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Concat_Name(void) {
   return "Concat";
