+++ src/Doc.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Doc_Name(void) {
   return "Doc";
