+++ src/Exception.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 #define EXCEPTION_TLS_KEY "__Exception"
