+++ src/File.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Stream_Name(void) {
   return "Stream";
