+++ src/Function.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Call_Name(void) {
   return "Call";
