+++ src/GC.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Mark_Name(void) {
   return "Mark";
