+++ src/Hash.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Hash_Name(void) {
   return "Hash";
