+++ src/Pointer.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Pointer_Name(void) {
   return "Pointer";
