+++ src/Push.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Push_Name(void) {
   return "Push";
