+++ src/Resize.c
@@ -1,4 +1,4 @@
-#include "Cello.h" 
+#include <Cello/Cello.h>
 
 static const char* Resize_Name(void) {
   return "Reserve";
