+++ src/Show.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Format_Name(void) {
   return "Format";
