+++ src/Start.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Start_Name(void) {
   return "Start";
