+++ src/String.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* C_Str_Name(void) {
   return "C_Str";
