+++ src/Thread.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Current_Name(void) {
   return "Current";
