+++ src/Tuple.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Tuple_Name(void) {
   return "Tuple";
