+++ src/Type.c
@@ -1,4 +1,4 @@
-#include "Cello.h"
+#include <Cello/Cello.h>
 
 static const char* Cast_Name(void) {
   return "Cast";
